files relocated from DAL to Models
