+++ Oef_Project_Destiny/Destiny_DAL/Destiny_DAL.csproj
@@ -63,6 +63,7 @@
     <Compile Include="Armor.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
+    <Compile Include="Basisklasse.cs" />
     <Compile Include="Character.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
@@ -72,8 +73,6 @@
     <Compile Include="CharacterSubklasse.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
-    <Compile Include="GeneralItems.cs" />
-    <Compile Include="OptiesUiterlijk.cs" />
     <Compile Include="Currency.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
@@ -105,7 +104,7 @@
     <Compile Include="Enemy.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
-    <Compile Include="fileOperations.cs" />
+    <Compile Include="FileOperations.cs" />
     <Compile Include="GranaatAbilityDamage.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
@@ -171,7 +170,6 @@
     <Compile Include="SubklassePerk.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
-    <Compile Include="User.cs" />
     <Compile Include="Wapen.cs">
       <DependentUpon>DestinyModel.tt</DependentUpon>
     </Compile>
@@ -204,12 +202,6 @@
       <DependentUpon>DestinyModel.edmx</DependentUpon>
       <LastGenOutput>DestinyModel.cs</LastGenOutput>
     </Content>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Destiny_Models\Destiny_Models.csproj">
-      <Project>{16F32F33-765A-4F53-A231-EAACE37ECB7C}</Project>
-      <Name>Destiny_Models</Name>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
