+++ Oef_Project_Destiny/Destiny_DAL/Partials/Account.cs
@@ -1,5 +1,4 @@
-﻿using Destiny_Models;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
