+++ Oef_Project_Destiny/Destiny_DAL/Partials/CharacterKlasse.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Destiny_Models;
+
 namespace Destiny_DAL
 {
    public partial class CharacterKlasse
