+++ Oef_Project_Destiny/Destiny_Models/Destiny_Models.csproj
@@ -55,14 +55,22 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Basisklasse.cs" />
+    <Compile Include="GeneralItems.cs" />
+    <Compile Include="OptiesUiterlijk.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SecurePassword.cs" />
+    <Compile Include="User.cs" />
     <Compile Include="Valideer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Destiny_DAL\Destiny_DAL.csproj">
+      <Project>{d17d8489-7bf6-4a47-adbb-c221288815ae}</Project>
+      <Name>Destiny_DAL</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
