+++ Oef_Project_Destiny/Destiny_Models/GeneralItems.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Destiny_DAL;
 
-namespace Destiny_DAL
+namespace Destiny_Models
 {
     public static class GeneralItems
     {
