+++ Oef_Project_Destiny/Project_Destiny_WPF/UserControls/ArmorControl.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using Destiny_DAL;
+using Destiny_Models;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
