+++ Oef_Project_Destiny/Project_Destiny_WPF/UserControls/CharacterChangeControl.xaml.cs
@@ -13,6 +13,8 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 using Destiny_DAL;
+using Destiny_Models;
+
 namespace Project_Destiny_WPF.UserControls
 {
     /// <summary>
@@ -134,8 +136,6 @@ namespace Project_Destiny_WPF.UserControls
                     }
                     
                 }
-
-
             }
             else
             {
