+++ Oef_Project_Destiny/Project_Destiny_WPF/UserControls/CharacterCreateControl.xaml.cs
@@ -13,6 +13,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Destiny_Models;
 
 namespace Project_Destiny_WPF.UserControls
 {
