+++ Oef_Project_Destiny/Project_Destiny_WPF/UserControls/LoggedInControl.xaml.cs
@@ -1,5 +1,4 @@
-﻿using Destiny_DAL;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -13,6 +12,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Destiny_Models;
 
 namespace Project_Destiny_WPF.UserControls
 {
