+++ Oef_Project_Destiny/Project_Destiny_WPF/UserControls/SpecialItemControl.xaml.cs
@@ -15,6 +15,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Destiny_Models;
 
 namespace Project_Destiny_WPF.UserControls
 {
